+++ src/core/utils.js
@@ -32,9 +32,19 @@ function createPromise() {
   return { promise, reject: failureResolve, fulfill: successResolve };
 }
 
+function stringToArrayBuffer(str) {
+  var buf = new ArrayBuffer(str.length * 2);
+  var bufView = new Uint16Array(buf);
+  for (var i = 0, strLen = str.length; i < strLen; i++) {
+    bufView[i] = str.charCodeAt(i);
+  }
+  return buf;
+}
+
 module.exports = {
   signPamFromParams,
   endsWith,
   createPromise,
   encodeString,
+  stringToArrayBuffer,
 };