+++ src/web/index.js
@@ -11,6 +11,7 @@ import { del, get, post, patch, getfile, postfile } from '../networking/modules/
 
 import WebCryptography from '../crypto/modules/web';
 import PubNubFile from '../file/modules/web';
+import { CryptoModule, LegacyCryptor, AesCbcCryptor } from '../crypto/modules/WebCryptoModule/webCryptoModule';
 
 function sendBeacon(url) {
   if (navigator && navigator.sendBeacon) {
@@ -21,10 +22,10 @@ function sendBeacon(url) {
 }
 
 export default class extends PubNubCore {
+  static CryptoModule = CryptoModule;
   constructor(setup) {
     // extract config.
     const { listenToBrowserNetworkEvents = true } = setup;
-
     setup.sdkFamily = 'Web';
     setup.networking = new Networking({
       del,
@@ -39,6 +40,16 @@ export default class extends PubNubCore {
 
     setup.PubNubFile = PubNubFile;
     setup.cryptography = new WebCryptography();
+
+    setup.initCryptoModule = (cryptoConfiguration) => {
+      return new CryptoModule({
+        default: new LegacyCryptor({
+          cipherKey: cryptoConfiguration.cipherKey,
+          useRandomIVs: cryptoConfiguration.useRandomIVs,
+        }),
+        cryptors: [new AesCbcCryptor({ cipherKey: cryptoConfiguration.cipherKey })],
+      });
+    };
 
     super(setup);