+++ test/contract/world.ts
@@ -4,7 +4,7 @@ import {
   World
 } from '@cucumber/cucumber';
 import PubNub from '../../lib/node/index.js';
-import { loadFixtureFile } from './utils';
+import { loadFixtureFile, getFilePath } from './utils';
 import * as http from 'http';
 
 interface State {
@@ -133,6 +133,9 @@ class PubnubWorld extends World{
       this.fileFixtures[name] = persona;
     }
     return this.fileFixtures[name];
+  }
+  getFilePath(filename) {
+    return getFilePath(filename);
   }
 }